backend/requirements.txt
anthropic>=0.40.0
PyGithub>=2.3.0
docker>=7.0.0
playwright>=1.40.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
